Projects/TensorflowFlaskAPIServer/requirements.txt
numpy==1.15.4
Flask_Cors==3.0.7
requests==2.21.0
imutils==0.5.2
Flask==1.0.2
Pillow==6.1.0
